requirements.txt
streamlit>=1.31
matplotlib>=3.8
pillow>=10.0
reportlab>=4.0
